pgoutput: refactor publication cache construction
Breaking this logic into a standalone helper should help expose the
behavior for testing.  Additionally, move the implementation under the
pg_publication catalog helpers, since it seems like it's inherent to the
publication settings (and it must match the behavior of the initial
sync, which doesn't seem to be controlled by the output plugin at all).