Bump the minimum supported Python version to 3.6.8
Python 3.2 is no longer tested by the buildfarm, and there are only a
handful of buildfarm animals running versions older than 3.6, which
itself went end-of-life in 2021. Python 3.6.8 is the default version
shipped in RHEL8, so that seems like a reasonable baseline for PG18.

Now that we use the Python Limited API as of 0793ab810, older versions
of Python should continue functioning for users of PL/Python in
particular, so soften the language from "required" to "supported".

Wording by Tom Lane. Several reviewers provided input for the choice of
cutoff (though the chosen cutoff is not necessarily what they all
preferred!).

Suggested-by: Tom Lane <tgl@sss.pgh.pa.us>
Reviewed-by: Christoph Berg <myon@debian.org>
Reviewed-by: Devrim Gündüz <devrim@gunduz.org>
Reviewed-by: Florents Tselai <florents.tselai@gmail.com>
Reviewed-by: Jelte Fennema-Nio <postgres@jeltef.nl>
Reviewed-by: Peter Eisentraut <peter@eisentraut.org>
Reviewed-by: Renan Alves Fonseca <renanfonseca@gmail.com>
Discussion: https://postgr.es/m/16098.1745079444%40sss.pgh.pa.us