Add pg_add_size_overflow() and friends
Commit 600086f47 added (several bespoke copies of) overflow-checked
size_t addition to libpq. Implement this in common/int.h, along with its
subtraction and multiplication counterparts, so that libpq can use that
instead.

pg_neg_size_overflow() is intentionally omitted. ssize_t is POSIX, not
C, and I'm not sure I should add SSIZE_MAX to win32_port.h for the sake
of a function that doesn't yet have a caller.